+++ src/pyoma2/algorithms/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from .base import BaseAlgorithm  # noqa
+from .base import BaseAlgorithm, MultiSetupData, SingleSetupData  # noqa
 from .data.run_params import FDDRunParams, SSIRunParams, pLSCFRunParams  # noqa
-from .fdd import EFDD, EFDD_MS, FDD, FDD_MS, FSDD  # noqa
+from .fdd import EFDD, EFDD_MS, FDD, FDD_MS, FSDD, FDDAlgorithm  # noqa
 from .plscf import pLSCF, pLSCF_MS  # noqa
 
 # from .ssi import SSIcov, SSIcov_MS, SSIdat, SSIdat_MS  # noqa